pyproject.toml
[project]
name = "polymarket-soccer-bot"
version = "0.1.0"
description = "Self-improving Polymarket soccer betting bot (Karpathy autoresearch style)"
requires-python = ">=3.10"
dependencies = [
    "py-clob-client>=0.17.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "eth-account>=0.11.0",
    "websockets>=12.0",
    "understatapi>=0.7.0",
    "soccerdata>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
